backend/requirements.txt
APScheduler==3.6.3
asgiref==3.7.2
astroid==2.14.1
atomicwrites==1.4.1
attrs==23.1.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
dill==0.3.6
Django==3.2
django-cors-headers==3.13.0
django-extra-fields==3.0.2
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
dnspython==2.2.1
flake8==3.9.2
flake8-docstrings==1.6.0
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.6.1
oauthlib==3.2.2
packaging==23.1
Pillow==9.3.0
platformdirs==3.0.0
pluggy==0.13.1
prettytable==3.6.0
psycopg2==2.9.9
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.3.1
PyJWT==2.8.0
pylint==2.16.1
pymongo==4.3.2
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytils==0.3
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.3.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
wcwidth==0.2.6
wrapt==1.14.1
